Get rid of third party packages, leave only multipart-kit from Vapor repository. Using self-written web-server for receiving Telegram Updates via Webhooks. Added target 'TestWebhooksLocally' will gives you possibility to test Webhooks locally. Improved Error logging. Now using only native NIO methods, without Vapor functional wrappers. Muted some deprecation warnings.

+++ Sources/Telegrammer/Helpers/TimeAmount+Helpers.swift
@@ -10,6 +10,6 @@ import NIO
 
 extension TimeAmount {
     var seconds: Int64 {
-        return Int(round(Double(self.nanoseconds) / 1_000_000_000.0))
+        return Int64(round(Double(self.nanoseconds) / 1_000_000_000.0))
     }
 }
